Prohibit giving a base to --logarithmic
Having an option with nargs="?" made it clunky, because to use it as a
plain flag without argument it has to be last option in command-line. At
same time, varying log base is not very useful.

Since frame delays after applying logarithmic time compression are still
subject to speed modifier, it makes sense to use fixed base e (ln)
logarithm, like it's done in IPBT and use option only as a flag to
enable it.

Signed-off-by: Vitaly Ostrosablin <tmp6154@yandex.ru>